Undo move of Cat from chisel3.util to chisel3
It is a very painful change for users with marginal benefit. We plan to
change the package to chisel in the not-so-distant future--that's a much
better opportunity to make this change.
