+++ integration-tests/src/test/scala/chiselTest/LFSRSpec.scala
@@ -4,7 +4,7 @@ package chiselTests.util.random
 
 import chisel3._
 import circt.stage.ChiselStage
-import chisel3.util.Counter
+import chisel3.util.{Cat, Counter}
 import chisel3.util.random._
 import chisel3.testers.{BasicTester, TesterDriver}
 import chiselTests.{ChiselFlatSpec, Utils}
