+++ src/test/scala/chiselTests/util/CatSpec.scala
@@ -1,9 +1,11 @@
 // SPDX-License-Identifier: Apache-2.0
 
-package chiselTests
+package chiselTests.util
 
 import chisel3._
+import chisel3.util.Cat
 import chisel3.experimental.noPrefix
+import chiselTests.ChiselFlatSpec
 import circt.stage.ChiselStage
 
 object CatSpec {
@@ -21,7 +23,7 @@ class CatSpec extends ChiselFlatSpec {
 
   import CatSpec._
 
-  behavior.of("chisel3.Cat")
+  behavior.of("chisel3.util.Cat")
 
   it should "not fail to elaborate a zero-element Vec" in {
